server/requirements.txt
fastapi[standard]
uvicorn[standard]
google-genai
google-cloud-storage
pandas
Pillow
python-multipart
python-dotenv